Changes to fix win animation for baki and update round state machine and certain attack meter gain on hit

+++ Project Strongest/Assets/FGC_MAKER/Prefabs/TrainingMenu_Prefab/Training_SubMenu/MoveList/UI_Movelist_SubMenu.cs
@@ -130,8 +130,8 @@ public class MoveListAttackInfo
 {
     public string AttackName;
     public string AttackInput;
-    public int meterRequirement;
-    public MoveListAttackInfo(string _name, string _input, int _m_requirement) 
+    public float meterRequirement;
+    public MoveListAttackInfo(string _name, string _input, float _m_requirement) 
     {
         AttackName = _name;
         AttackInput = _input;
