+++ Project Strongest/Assets/FGC_MAKER/Scripts/Character_Code/Character_AttackCode/Attack_Versions/Attack_BasicSpecialMove.cs
@@ -67,7 +67,7 @@ public class Attack_BasicSpecialMove : Attack_Special_Base, IAttackFunctionality
     {
         string attackName = property._attackName;
         string specialMoveInput = attackInput[0].attackString;
-        int _meterRequirement = property._meterRequirement;
+        float _meterRequirement = property._meterRequirement;
         MoveListAttackInfo newMoveListAttackInfo = new MoveListAttackInfo(attackName, specialMoveInput, _meterRequirement);
 
         return newMoveListAttackInfo;
