+++ Project Strongest/Assets/FGC_MAKER/Scripts/Character_Code/Character_AttackCode/Attack_Versions/Attack_ThrowBase.cs
@@ -43,7 +43,7 @@ public class Attack_ThrowBase : ThrowActionBase , IAttackFunctionality
     {
         string attackName = _attackInput._correctInput[0].property._attackName;
         string specialMoveInput = _attackInput._correctInput[0]._correctSequence;
-        int _meterRequirement = _attackInput._correctInput[0].property._meterRequirement;
+        float _meterRequirement = _attackInput._correctInput[0].property._meterRequirement;
         MoveListAttackInfo newMoveListAttackInfo = new MoveListAttackInfo(attackName, specialMoveInput, _meterRequirement);
 
         return newMoveListAttackInfo;
